activity labels and toggle working
